Fixed clippy cast warnings for viewer & client, simplified db_list.rs to use ? operator and HOF instead of match expressions.

+++ smol_db_server/src/main.rs
@@ -89,7 +89,7 @@ fn main() {
                 .cache
                 .read()
                 .unwrap()
-                .len() as u32;
+                .len();
             let msg = format!(
                 "Slept {} caches, {} caches remain in cache.",
                 invalidated_caches, number_of_caches_remaining
